refactor: combine RBMDS and RBMNcDS into one class (fixes #67)
