+++ swvo/io/RBMDataSet/__init__.py
@@ -20,8 +20,6 @@ from swvo.io.RBMDataSet.custom_enums import (
     SatelliteLiteral as SatelliteLiteral,
     VariableLiteral as VariableLiteral,
 )
-from swvo.io.RBMDataSet.RBMDataSetManager import RBMDataSetManager as RBMDataSetManager
 from swvo.io.RBMDataSet.interp_functions import TargetType as TargetType
 from swvo.io.RBMDataSet.scripts.create_RBSP_line_data import create_RBSP_line_data as create_RBSP_line_data
 from swvo.io.RBMDataSet.RBMDataSet import RBMDataSet as RBMDataSet
-from swvo.io.RBMDataSet.RBMNcDataSet import RBMNcDataSet as RBMNcDataSet
